refactor: move files to separate packages
