+++ dash/dip6/sorted_validator_list_test.go
@@ -1,0 +1,45 @@
+package dip6
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/tendermint/tendermint/dash/mock"
+)
+
+// Test_sortableValidatorList_Index checks if the sortableValidatorList.Index() works correctly
+func Test_sortedValidatorList_Index(t *testing.T) {
+	tests := []struct {
+		name   string
+		list   sortedValidatorList
+		search sortableValidator
+		want   int
+	}{{
+		name:   "miss",
+		list:   newSortedValidatorList(mock.Validators(5), mock.QuorumHash(0)),
+		search: newSortableValidator(mock.Validator(10), mock.QuorumHash(0)),
+		want:   -1,
+	}, {
+		name:   "i=0",
+		list:   newSortedValidatorList(mock.Validators(500), mock.QuorumHash(0)),
+		search: newSortableValidator(mock.Validator(0), mock.QuorumHash(0)),
+		want:   0,
+	}, {
+		name:   "i=4",
+		list:   newSortedValidatorList(mock.Validators(500), mock.QuorumHash(2054231)),
+		search: newSortableValidator(mock.Validator(4), mock.QuorumHash(2054231)),
+		want:   4,
+	}}
+	// nolint:scopelint
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got := tt.list.Index(tt.search)
+			if tt.want < 0 {
+				assert.Less(t, got, 0)
+			} else {
+				assert.EqualValues(t, tt.search, tt.list[got])
+			}
+
+		})
+	}
+}
