+++ dash/mock/mock_switch.go
@@ -1,4 +1,4 @@
-package dash
+package mock
 
 import (
 	"fmt"
@@ -11,48 +11,48 @@ import (
 
 // MOCK SWITCH //
 const (
-	OpDialMany historyOperation = "dialMany"
-	OpStopOne  historyOperation = "stopOne"
+	OpDialMany SwitchHistoryOperation = "dialMany"
+	OpStopOne  SwitchHistoryOperation = "stopOne"
 )
 
-type historyOperation string
+type SwitchHistoryOperation string
 
-// mockSwitchHistoryEvent is a log of dial and stop operations executed by the MockSwitch
-type mockSwitchHistoryEvent struct {
+// SwitchHistoryEvent is a log of dial and stop operations executed by the MockSwitch
+type SwitchHistoryEvent struct {
 	Timestamp time.Time
-	Operation historyOperation // OpDialMany, OpStopOne
+	Operation SwitchHistoryOperation // OpDialMany, OpStopOne
 	Params    []string
 	Comment   string
 }
 
-// MockSwitch implements dash.ISwitch. It sends event about DialPeersAsync() and StopPeerGracefully() calls
+// Switch implements dash.ISwitch. It sends event about DialPeersAsync() and StopPeerGracefully() calls
 // to HistoryChan and stores them in History
-type MockSwitch struct {
+type Switch struct {
 	PeerSet         *p2p.PeerSet
 	PersistentPeers map[string]bool
-	History         []mockSwitchHistoryEvent
-	HistoryChan     chan mockSwitchHistoryEvent
+	History         []SwitchHistoryEvent
+	HistoryChan     chan SwitchHistoryEvent
 }
 
 // NewMockSwitch creates a new mock Switch
-func NewMockSwitch() *MockSwitch {
+func NewMockSwitch() *Switch {
 
-	isw := &MockSwitch{
+	isw := &Switch{
 		PeerSet:         p2p.NewPeerSet(),
 		PersistentPeers: map[string]bool{},
-		History:         []mockSwitchHistoryEvent{},
-		HistoryChan:     make(chan mockSwitchHistoryEvent, 1000),
+		History:         []SwitchHistoryEvent{},
+		HistoryChan:     make(chan SwitchHistoryEvent, 1000),
 	}
 	return isw
 }
 
 // Peers implements ISwitch by returning sw.PeerSet
-func (sw *MockSwitch) Peers() p2p.IPeerSet {
+func (sw *Switch) Peers() p2p.IPeerSet {
 	return sw.PeerSet
 }
 
 // AddPersistentPeers implements ISwitch by marking provided addresses as persistent
-func (sw *MockSwitch) AddPersistentPeers(addrs []string) error {
+func (sw *Switch) AddPersistentPeers(addrs []string) error {
 	for _, addr := range addrs {
 		addr = strings.TrimPrefix(addr, "tcp://")
 		sw.PersistentPeers[addr] = true
@@ -62,9 +62,9 @@ func (sw *MockSwitch) AddPersistentPeers(addrs []string) error {
 
 // RemovePersistentPeer implements ISwitch. It checks if the addr is persistent, and
 // marks it as non-persistent if needed.
-func (sw *MockSwitch) RemovePersistentPeer(addr string) error {
+func (sw *Switch) RemovePersistentPeer(addr string) error {
 	addr = strings.TrimPrefix(addr, "tcp://")
-	if sw.PersistentPeers[addr] != true {
+	if !sw.PersistentPeers[addr] {
 		return fmt.Errorf("peer is not persisitent, addr=%s", addr)
 	}
 
@@ -74,7 +74,7 @@ func (sw *MockSwitch) RemovePersistentPeer(addr string) error {
 
 // DialPeersAsync implements ISwitch. It emulates connecting to provided addresses
 // and adds them as peers and emits history event OpDialMany.
-func (sw *MockSwitch) DialPeersAsync(addrs []string) error {
+func (sw *Switch) DialPeersAsync(addrs []string) error {
 	for _, addr := range addrs {
 		peer := &mocks.Peer{}
 		parsed, _ := p2p.ParseNodeAddress(addr)
@@ -91,20 +91,20 @@ func (sw *MockSwitch) DialPeersAsync(addrs []string) error {
 
 // IsDialingOrExistingAddress implements ISwitch. It checks if provided peer has been dialed
 // before.
-func (sw *MockSwitch) IsDialingOrExistingAddress(addr *p2p.NetAddress) bool {
+func (sw *Switch) IsDialingOrExistingAddress(addr *p2p.NetAddress) bool {
 	return sw.PeerSet.Has(addr.ID)
 }
 
 // StopPeerGracefully implements iSwitch. It removes the peer from Peers() and emits history
 // event OpStopOne.
-func (sw *MockSwitch) StopPeerGracefully(peer p2p.Peer) {
+func (sw *Switch) StopPeerGracefully(peer p2p.Peer) {
 	sw.PeerSet.Remove(peer)
 	sw.history(OpStopOne, peer.String())
 }
 
 // history adds info about an operation to sw.History and sends it to sw.HistoryChan
-func (sw *MockSwitch) history(op historyOperation, args ...string) {
-	event := mockSwitchHistoryEvent{
+func (sw *Switch) history(op SwitchHistoryOperation, args ...string) {
+	event := SwitchHistoryEvent{
 		Timestamp: time.Now(),
 		Operation: op,
 		Params:    args,
