+++ dash/validator_conn_executor.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 
 	"github.com/tendermint/tendermint/crypto"
+	"github.com/tendermint/tendermint/dash/dip6"
 	tmbytes "github.com/tendermint/tendermint/libs/bytes"
 	"github.com/tendermint/tendermint/libs/log"
 	"github.com/tendermint/tendermint/libs/service"
@@ -199,7 +200,7 @@ func (vc *ValidatorConnExecutor) selectValidators() (validatorMap, error) {
 		return validatorMap{}, fmt.Errorf("current node is not member of active validator set")
 	}
 
-	selectedValidators, err := SelectValidatorsDIP6(activeValidators.values(), me, vc.quorumHash)
+	selectedValidators, err := dip6.SelectValidatorsDIP6(activeValidators.values(), me, vc.quorumHash)
 	if err != nil {
 		return validatorMap{}, err
 	}
